hardens my pp

+++ code/game/objects/items/discoball.dm
@@ -1,17 +1,21 @@
 /obj/item/etherealballdeployer
 	name = "Portable Ethereal Disco Ball"
-	desc = "Press the button for a deployment of PARTY!"
-	icon = 'icons/mob/human_parts.dmi'
-	icon_state = "ethereal_head_m"
+	desc = "Press the button for a deployment of slightly-unethical PARTY!"
+	icon = 'icons/obj/eth_disco.dmi'
+	icon_state = "ethdisco"
 
 /obj/item/etherealballdeployer/attack_self(mob/living/carbon/user)
+	var/mob/living/carbon/human/coolperson = user
+	if(!(coolperson.ckey == "qustinnus" || coolperson.ckey == "mrdoombringer"))
+		to_chat(user, "<span class='notice'>Hello buddy, sorry, only cool people can turn the Ethereal Ball 3000 on or off, you can feed it or give it water, though!</span>")
+		return
 	new /obj/structure/etherealball(user.loc)
 	qdel(src)
 
 /obj/structure/etherealball
 	name = "Ethereal Disco Ball"
 	desc = "The ethics of this discoball are questionable. Be sure to feed it snacks or else it might turn off!"
-	icon = 'icons/obj/objects.dmi'
+	icon = 'icons/obj/eth_disco.dmi'
 	icon_state = "ethdisco_head_0"
 	anchored = TRUE
 	density = TRUE
@@ -25,8 +29,6 @@
 
 /obj/structure/etherealball/attack_hand(mob/living/carbon/human/user)
 	. = ..()
-	if(!ishuman(user))
-		return //Bish we only play human
 	var/mob/living/carbon/human/coolperson = user
 	if(!(coolperson.ckey == "qustinnus" || coolperson.ckey == "mrdoombringer"))
 		to_chat(user, "<span class='notice'>Hello buddy, sorry, only cool people can turn the Ethereal Ball 3000 on or off, you can feed it or give it water, though!</span>")
@@ -37,6 +39,17 @@
 	else
 		TurnOn()
 		to_chat(user, "<span class='notice'>You turn the disco ball on!</span>")
+
+/obj/structure/etherealball/AltClick(mob/living/carbon/human/user)
+	. = ..()
+	if(!ishuman(user))
+		return //Bish we only play human
+	var/mob/living/carbon/human/coolperson = user
+	if(!(coolperson.ckey == "qustinnus" || coolperson.ckey == "mrdoombringer"))
+		to_chat(user, "<span class='notice'>Hello buddy, sorry, only cool people can pack up the Ethereal Ball 3000!</span>")
+		return
+	new /obj/item/etherealballdeployer(user.loc)
+	qdel(src)
 
 /obj/structure/etherealball/proc/TurnOn()
 	TurnedOn = TRUE //Same
